consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["ilyakochankov <ilyakochankov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
Faker = "^13.11.0"
pymongo = "^4.1.1"
environs = "^9.5.0"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]
black = {extras = ["d"], version = "^22.3.0"}
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
